Query engine now use API suggested by @angru, while remaining compatible with django-like queries
